Slightly clean Participant#build_sample_days
I was going to change it to start with an hour
from now, but if you're signing up at like 10 PM
I don't want it to be all "you must get texts
for the next like 8 hours" because that would be
annoying.

We can change this if we allow skipping entire
days someday.